refactor(common): migrate monetary primitives
