+++ src/libs/portfolio-common/portfolio_common/transaction_fee_components.py
@@ -1,7 +1,7 @@
 from decimal import Decimal
 from typing import Mapping
 
-from portfolio_common.decimal_amounts import decimal_or_none
+from portfolio_common.domain.decimal_amount import decimal_or_none
 
 TRANSACTION_FEE_COMPONENT_FIELDS = (
     "brokerage",
